File Menu - New (IN PROGRESS)
Some very in progress work on fixing the New file menu item. This batch
of code just destroys the instance of the music file, disables the
playback buttons, and resets the music file label and time indicators on
the GUI. It will raise some exceptions, because I still need to work on
resetting the Sim, Timeline, CTL file, beat marks, and color pallet.