Fixed: File->New
Still not able remove instance of loaded control file. Other than that,
it looks to reset properly.
I decided to disable the ability to click new when first launching the
problem. This is due to the fact that clicking it caused major
exceptions be thrown since nothing was loaded in to begin with.