Fixed ColorPalette
The ColorPalette is "fixed" in the sense that I had to remove the
ColorPicker from the ContextMenu in order for the "Custom Colors" link
to work. Now you have to select a color from the ColorPicker and then
right-click the rectangle and select ChangeColor. I submitted a bug
report to JavaFX in the hopes that maybe they could help me find a work
around to achieve the previously desired effect, but for now this will
do.