Lag time fix started
Saving a ctl file adds in a correct time adjustment, though not
necessarily from reading the lag time def text file. Reason for this is
that the lookups depend on reverseLookUpFCW/FCWLib.java methods, and
these, once again, aren't implemented 100% correctly. Also added in a
zip change so that opening up (and saving) as a zip works some of the
time, with odd edge cases and quirks, as per usual.