Second stateful timeline
I wrote a second timeline which contains all the states and not just
changes. They both run parallel and don't influence each other.

With the stateful timeline you can click on pause during reflection on
earth and see all the state of the sliders or you can make a new
choreography and also see all the states.

The simulation is still stateless -> but could be changed to stateful
now....